ユーザー機能の作成

+++ app/controllers/articles_controller.rb
@@ -1,4 +1,6 @@
 class ArticlesController < ApplicationController
+  before_action :move_to_index, except: :index
+
   def index
     @articles = Article.order("created_at DESC")
   end
@@ -15,22 +17,28 @@ class ArticlesController < ApplicationController
 
   def destroy
     article = Article.find(params[:id])
-    article.destroy
+    article.destroy if article.user_id == current_user.id # 自分の記事でない場合、削除しない
     redirect_to :root
   end
 
   def edit
     @article = Article.find(params[:id])
+    redirect_to :root unless @article.user_id == current_user.id # 自分の記事でない場合、ルートパスにリダイレクト
   end
 
   def update
     article = Article.find(params[:id])
-    article.update(article_params)
+    article.update(article_params) if article.user_id == current_user.id # 自分の記事でない場合、更新しない
     redirect_to :root
   end
 
   private
   def article_params
-    params.require(:article).permit( :title, :text )
+    params.require(:article).permit( :title, :text ).merge(user_id: current_user.id)
+  end
+
+  # ログインしていなかったらindexに飛ばす
+  def move_to_index
+    redirect_to action: :index unless user_signed_in?
   end
 end
